Renamed the release version of fannkuch, to avoid error in start_test -performance.
Also updated READMEs and changed .perfexecopts and an error message in .chpl
from --N to --n.

Before the rename, start_test -performance reported the error:
  Duplicate performance data filenames (fannkuch-redux.dat)

perhaps due to the seeming conflict with
  studies/shootout/fannkuch-redux/kbrady/fannkuch-redux.chpl

In this case, start_test approximates -- the actual .dat file
for the fannkuch release version is given it its .perfexecopts:
  fannkuch-redux-release
This commit does not change that.

Naming the test 'fannkuchredux' in the release directory seems consistent with
the existing naming scheme in that directory, e.g. chameneosredux.chpl.
Nonetheless any adjustments are welcome.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23003 3a8e244f-b0f2-452b-bcba-4c88e055c3ca