Count running tasks in the module code, not the runtime tasking layers.
[Reviewed by Sung.]

We use the count of currently running tasks, as well as other things,
when we're computing how many tasks to create for a forall-stmt if
dataParIgnoreRunningTasks=false.  In the past the runtime tasking layer
implementations have maintained this count of running tasks, and it
reflected tasks running anywhere on the top-level locale (system node).
But with the advent of hierarchical locales, we really need to know the
number of tasks running on the particular place (locale or sublocale)
where the current task is executing.  The runtime cannot do that for us,
because it doesn't know about locale hierarchy.  So in this commit, we
move maintenance of the running task count into the module code.

Note that the current commit isn't the last word here.  If we have
CHPL_LOCALE_MODEL=numa we sometimes count a task on the wrong
(sub)locale.  This happens because for some constructs, the parent task
is on the top-level locale but the child tasks will be run on the NUMA
sublocales, and the children switch to their sublocales within their
task bodies.  But we increment and decrement the running task count for
the children in the parent, so it's the parent locale's counter that
reflects the existence of the child tasks.  We actually should be
adjusting the running task counters on the child locales instead, but
that will require compiler changes or a reworking of how parallelism is
initiated, in the locale model module code.  This only makes us off by
one in each child, and subsequent nested parallelism will update the
counter on the right sublocale, so for now we'll live with this.  (It's
still better than what we had before, and since for the numa locale
model we set dataParIgnoreRunningTasks=true by default, most users won't
see this problem anyway.)

The specific file changes are as follows.

modules/internal/ChapelLocale.chpl
  Add a running task counter to the locale class, along with procedures
  to set it, add to it, subtract from it, and retrieve its value.
  Modify the existing runningTasks() method to return this counter.

  Add exported procedures to increment and decrement the running task
  count in the base locale class.  These are intended for the runtime to
  use, to adjust the count for tasks created directly by it, without
  going through the module code.  We expect that, over time, the need
  for these procedures should diminish and eventually disappear.  Note
  that these procedures are not safe to call until after the modules
  have been initialized, and since they are a stopgap solution anyway,
  the responsibility for calling them only when it is safe to do so is
  placed on the calling code in the runtime.

modules/internal/localeModels/flat/LocaleModel.chpl
modules/internal/localeModels/numa/LocaleModel.chpl
  Zero the local running task count after setting locale initialization,
  because the count may have been left containing a meaningless value if
  there were tasking constructs in the locale initialization code.
  (Notably, initialization would have zeroed the count, but if that was
  done in a tasking construct, completion of the construct will have
  decremented it to -1.)

modules/internal/ChapelBase.chpl
  Increment the running task count in _upEndCount(), which the parent
  calls once for each task it will create for a parallel construct.  In
  addition, count these tasks in a new member of the EndCount class, and
  use that count to reduce the running task count in _waitEndCount(),
  after all the child tasks have completed.  This isn't ideal (see the
  general comment for the commit, above): with the numa locale model I
  believe there are cases in which the child moves to a sublocale after
  it is created, and if that happens, we manipulate the counter in the
  parent's locale on behalf of the child tasks, rather than that in the
  child's locale.  We'll need to revisit this at some point, but I
  believe for now this will give us the same behavior as we had in the
  past, when the runtime was counting the running tasks for the locale
  as a whole.

modules-minimal/internal/ChapelStandard.chpl
modules-minimal/internal/ChapelLocale.chpl
  Add vacuous versions of the procedures the modules export to the
  runtime, to increment and decrement the running task count for tasks
  the runtime creates itself.  Without this, we get link errors when we
  use --minimal-modules.

runtime/include/chpl-tasks.h
  Add declarations for chpl_taskRunningCnt{Inc,Dec}(), which runtime
  code can use to adjust the count for tasks created directly by it,
  without going through the module code.

runtime/src/tasks/fifo/tasks-fifo.c
third-party/qthread/qthread-1.10/src/interfaces/chapel/tasks-qthreads.h
third-party/qthread/qthread-1.10/src/interfaces/chapel/tasks-qthreads.c
runtime/include/tasks/massivethreads/tasks-massivethreads.h
runtime/src/tasks/massivethreads/tasks-massivethreads.c
  In the chpl_task_stdModulesInitialized() function called from the
  startup code after module initialization is complete, set a flag
  saying that it's now safe to manipulate the running task counters via
  the modules-provided interface, and increment the count for the user
  main task.

  Add a wrapper for task bodies started by chpl_task_startMovedTask(),
  and increment and decrement the running task count for such tasks in
  it.  Determine whether or not to adjust the counts for such tasks at
  the time they are enqueued, rather than when they run, in order to
  prevent the possibility of unbalanced increment/decrement pairs if
  module initialization ever creates long-lived tasks whose lifetimes
  span the point at which it becomes safe to manipulate the counters.

  Replace the body of chpl_task_getRunningTasks() with an internal error
  call, because we don't expect to have it called any more.  (I'll
  adjust the tasking interface to remove it after the other tasking
  layers have been updated.)

  In chpl_task_getNumBlockedTasks() for tasks=fifo and tasks=qthreads,
  return 0 rather than -1 if we're not counting blocked tasks.  A -1
  value messes up the running task computation in ChapelLocale.chpl
  quite badly!


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23004 3a8e244f-b0f2-452b-bcba-4c88e055c3ca