Created types/records/const-checking and moved one test there.
I'd like a directory for tests related to constness checking for records
that's separate from types/records/vass, so I created one.

I primed it with a test moved from types/records/vass,
which I renamed, modified, and created .bad:
  vass/constnessOfThisInMethods.chpl -> const-checking/this-in-method.chpl

I changed the test by leaving only a single error it in,
so the .good does not depend whether that error or both errors
are reported in a single compiler run.

The other error - assignment to a const variable of a record type -
will be tested in a separate uncoming test.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23008 3a8e244f-b0f2-452b-bcba-4c88e055c3ca