Added a few tests related to record const checking - r22963.
When the futures here are solved, we should be able to revert
most of the changes to made to tests in r22963.

TODO: add a relevant subset of similar tests for class fields.

BTW the discussion of LocaleModel.chpl and the callStackSize field
in the commit message for r22963 should not be there. That's because
I worked around that issue by allowing the functions called from
constructors and initialize() to tread 'const' fields as 'var' fields.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23009 3a8e244f-b0f2-452b-bcba-4c88e055c3ca