Fix duplicate symbol error observed with --llvm
These tests define ntrials both in C and Chapel...  and with different types
(Chapel int is 64 bits, C int is 32).  I've only seen the error with --llvm.
This patch makes the C variables 'static' so that they are not available to
conflict with the Chapel variable.

Future work: I will check that the LLVM backend treats non-exported symbols as
'static'.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23011 3a8e244f-b0f2-452b-bcba-4c88e055c3ca