Do not force wide strings to be passed by ref
[reviewed by sung] 

argMustUseCPtr forces all records and unions to be passed by ref in the
generated code. Since wide strings are a string wrapped with a record this
meant that wide strings were being passed by ref regardless of the actual
intent. 

This caused problems for types/tuples/deitz/test_string_and_tuples.chpl under
numa where a wide string was being passed by ref when it should have been
passed by value. The reason this wasn't a problem for --no-local, gasnet, etc
was because even though wide strings were passed by reference before they were
passed a temp copy is made and that is passed by reference. So for numa we had
something like: 

widestring s = ... ;
foo(&s);

and for --no-local we had:

widestring s = ...;
widestring tmp = s;
foo(&tmp);

so the problem that wide strings were being passed by ref was hidden by the
fact that a temp copy was being passed by ref for --no-local for but numa the
actual string was passed by ref. 

This patch fixes the current regression with
types/tuples/deitz/test_string_and_tuples.chpl under numa runs (note that is
only fails comm none with numa because for gasnet a temp copy of the string
was passed in. 

The function argMustUseCPtr may need to be looked at for other wide types as
well but for now it's only being changed for wide strings. 



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23014 3a8e244f-b0f2-452b-bcba-4c88e055c3ca