Improve while loop const checking
[reviewed by vass]

In r21919, Vass introduced a warning for 'while' loops over
certain constants.  This commit relaxes that warning in the
event that the const is local to the while loop (in which
case there's no reason to assume it wouldn't be a completely
different value from one iteration to the next).  In
particular, since do...while loops can refer to variables
declared within the loop's scope, the following idiom
shouldn't generate a warning by default:

  do {
    const x = ...
  } while (x);

I implemented this by walking from x's defPoint up the
parentExpr stack until I either hit NULL or the loop
block itself.

Also added a test to lock in this new behavior (as will
my rewrite of chameneos-redux).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23015 3a8e244f-b0f2-452b-bcba-4c88e055c3ca