Normalize behavior of cast to Boolean types
[reviewed by Tom]

Before this patch, casting int/uint to a Boolean type would sometimes give a
true value if the int/uint was nonzero, but sometimes truncate the int/uint
to the size of the Boolean.

The included test case demonstrates this problem (and should always print out
'true').  Code using default-width Bools worked OK because of the C99 _Bool
rules, but any cast to non-default width Boolean type or any cast to a Boolean
type with the LLVM backend did not work properly.

The simple fix in expr.cpp was to:
 1) modify a call to codegenCast to pass the Chapel type instead of
    type->symbol->cname, so that we can call is_bool_type on it.
 2) modify the codegenCast that takes in a Chapel type to handle Boolean
    types specially.


Verified with CHPL_LLVM=none make check and checked that the new test works
with and without --llvm; and completed full testing for local/C backend
and --llvm.

Future Work: update the spec to describe the behavior of casts to Bool.

Alternative Implementation: It might be possible to hande these cases
using Chapel's user-defined casting mechanism and to define a generic
cast-from-integral-to-bool in the modules.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23018 3a8e244f-b0f2-452b-bcba-4c88e055c3ca