Change the verification messages to improve readability
The performance verification messages could be confusing when using reject
lines, so now the checks in the form of:

Checking for /2098 solutions found/ on line 1... SUCCESS.
Checking for absence of /1 2 3 3 3/ on line 23... FAILURE.

If any message outputs 'FAILURE', verification for that test will have
failed. This should make the results of the verify checks much more
transparent and easier to understand without opening the .perfkeys.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23021 3a8e244f-b0f2-452b-bcba-4c88e055c3ca