Remove an unneeded verbose comm message.
[Reviewed conversationally by Sung, Jun, and Brad.]

The verbose comm message printed on the target side of a fast remote
fork causes an error in gcc 4.6 (I believe it is) due to our ignoring
the return value from write(2).  All potential fixes for this problem
are unattractive for one reason or another, so here we simply remove
the code that prints the message.  It is problematic because it is
running in a signal handler and is highly limited in what it can call,
we still have verbose comm for the same transaction on the initiating
side, and there is no test that checks for this message.  Removing it
seems the best solution.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23029 3a8e244f-b0f2-452b-bcba-4c88e055c3ca