Don't ignore the return value from fscanf(3).
[Patch by Jun, reviewed by Greg.  Committed by Greg just before code
cutoff so we have it for the release, since it's very late at night for
Jun.]

A new version of gcc (4.6, I believe) produces error messages for code
that ignores the value returned by a number of Linux- and C-standard
functions, including fscanf(3).  With this change, we no longer ignore
the return value from fscanf(3).


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23030 3a8e244f-b0f2-452b-bcba-4c88e055c3ca