Enable the slurm launcher for cray machines that have slurm
[reviewed by brad]

This just updates the chplenv launcher script to use slurm if it is available.
Currently we check if aprun and srun are available. And if only one is
available go with that one, and if both are available we default to none. Note
that we're not 100% sure if you can have both available at the same time so we
default to no launcher if both are available but we might update this shortly
to pick one over the other if they are both available. As in right now if aprun
is available and slurm is not we default to aprun, but if somehow aprun and
slurm are available we leave it up to the user to pick what launcher the want
and do not set a default launcher. 



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23031 3a8e244f-b0f2-452b-bcba-4c88e055c3ca