Enable -fstrict-overflow and quiet related warnings.
After GCC 4.2, optimizations related to -fstrict-overflow are enabled by
default when -O2 or -O3 is thrown.  Code in rangeBase.these() for a bounded
range cause -fstrict-overflow warnings to be issued, to we now toss
-Wno-strict-overflow to suppress these warnings.

This has the unfortunate side-effect of silencing warnings that the user might
be interested in.  Changing the form of the iterator to avoid generating those
warnings is left for future work.

The advantage is that faster execution of for loops is obtained.

[reviewed by sungeun, bradc]
[tested on overflow.chpl]
[tested on time_iterate.chpl -performance]


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23039 3a8e244f-b0f2-452b-bcba-4c88e055c3ca