Alter chameneos's getComplement function in the release.
It was also discovered that we were computing this arithmetically when
we should have been using a select statement or an if/else branch.  I have
changed the release version to use a select statement and updated
README.features to highlight this usage.  After code freeze, I will
investigate which is more performant.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23041 3a8e244f-b0f2-452b-bcba-4c88e055c3ca