Adapt call stack size setting logic to Cygwin limitations.
[Tested (thanks!) and reviewed by Brad.]

In Cygwin we can't do setrlimit(RLIMIT_STACK), and the stack limit seems
to be 2 MiB.  Adapt the stack size setting logic to these limitations.

In chpl-tasks.c I made both the default and max stack size functionally
defined, and used those to set the returned value for the default stack
size.  I also improved the messages we print when there are problems.

In threads-pthreads.c I just added a comment where the setrlimit was
already commented out.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23045 3a8e244f-b0f2-452b-bcba-4c88e055c3ca