Further refactoring/reorganization of chameneos
This is a pretty significant rewrite compared to my previous ones,
mostly from an OOP/design perspective.

It seemed weird to me to have the chameneos calling routines like CAS
on the MeetingPlace's state value -- which feels like a fairly
encoded/encrypted value whose semantics should be contained within the
meeting place, so I started by moving any routines dealing with the
state into MeetingPlace and out of Chameneos.  For example, I added
methods which convert a (#meetings, chameneos ID) <-> a state value to
keep the encoding of the stack a bit more of a black box.

As part of this, it seemed to make more sense to have the meeting
place host the meetings rather than the chameneos looking over into
the meeting place all the time.  So I moved the parallel loop over the
chameneos population itself into a new method
MeetingPlace.hostMeetings().

Then I put the two actions a chemeneos can take during a meeting --
waiting for it to end or running the meeting -- as methods within
the chameneos itself.

This overall reflects better organization and design to me without
really changing the algorithm at all -- just how it's packaged.  From
what I've seen on my desktop, it doesn't seem to really impact
performance much.  I'm not quite getting my record timings, but close
(and still better than last night's).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23047 3a8e244f-b0f2-452b-bcba-4c88e055c3ca