Bumping memory limit for this test
[trivial, not reviewed]

We've been seeing an uptick in mallocOutOfMemory compiler segfaults
lately, and I think the reason is that the sub_test script for this
test artificially dials the amount of memory available down.  This
is intended for use by the execution, but of course it affects
compilation as well.  So my theory is that our compiler has started
using more memory recently (or is more on the cusp?) and that this
is causing the segfaults.

I'm not sure why these would appear as segfaults rather than 
out of memory messages, though.  I tried to reproduce, but was
unable to.  Might be worth a look.

Anyway, I'm taking the approach of bumping up the amount of memory
in the limit statement.  

The better way to fix this issue would be to limit the memory only for
the executable.  I had hoped to switch over to the .execenv
capability, but this only permits setting environment variables, not
executing limit commands (so far as I can tell), so that wasn't an
option.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23053 3a8e244f-b0f2-452b-bcba-4c88e055c3ca