Subject: clarified the language in the spec w.r.t. the 'inline' keyword
[reviewed by Brad]

Brad's reminder: The last time we discussed it, we decided to make it
an imperative assertion "This needs to be inlined" rather than a hint.

I removed the "for improved performance" clause because it implies an
option not to inline, if the compiler decides that the performance
would degrade because of inlining.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23062 3a8e244f-b0f2-452b-bcba-4c88e055c3ca