Don't run module code that needs 'here' before it can tell where that is!
[Reviewed by Sung.]

In r23004 I added a call to module code to increment the running task
count on 'here' for moved tasks.  Unfortunately I added that call just
before to the point at which we store the task local data that allows
the module code to tell where 'here' is in the first place.  The result
was an error message like this one:

  error: halt reached - sublocale child index out of bounds (-469761928)

for studies/sort/radix with comm=none, tasks=fifo, localeModel=flat, due
to the reference to uninitialized data.

This commit moves the increment call to a point after the necessary task
local data has been set.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23066 3a8e244f-b0f2-452b-bcba-4c88e055c3ca