Update operator precedence in spec and quick-reference.
Rearranged the entries in the operator precedence table in the spec,
to reflect the changes in precedence Michael made in the parser.  I
cross-checked with the code in chapel.ypp to verify accuracy.

Basically, the shift and bitwise boolean operators were moved above
above addition, and range construction was moved above comparison.  I
also noticed that the table failed to mention "align" or "atomic", so
I added these.

An explanatory paragraph was added (pulled from Michael's commit
message) to justify the reordering of shifts and bitwise operations
w.r.t. addition and subtraction.  Since C doesn't have a range
creation operator, there was nothing to say about the other move.

These changes are important, and bear mentioning in the CHANGES file
(in case that has not already been done).

The changes were propagated verbatim to the quick reference card, and
an updated PDF generated, ready for checkin.

[reviewer: lydia]
[test-built spec and checked formatting]


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23072 3a8e244f-b0f2-452b-bcba-4c88e055c3ca