Update .good/.bad files and future for comm count regression due to r23004
The endCountTest and synsingle test have an additional get and put due
to the new task counting changes, because they both spawn begin
statements on non-0 locales, and thus must up the task count for the
implicit sync statement around main.

I added a new performance future (taskCount.chpl) for these extra
comms.  It's the same test as endCountTest, so I added a main() to
that endCount.chpl and used it in taskCount.chpl.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23075 3a8e244f-b0f2-452b-bcba-4c88e055c3ca