Removing the "AGREEMENT" file
This file was, historically, a way for us to cover our butts in
certain respects as requested by Cray legal at the time.  Cray legal
has now requested we remove it (so as not to imply that our license is
something other than strictly BSD), so we'll work on covering our
butts in other ways, to the extent that we still feel it's necessary.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23076 3a8e244f-b0f2-452b-bcba-4c88e055c3ca