Retiring ra-dist.chpl performance test
This test doesn't add anything over the release version of RA and has
the downside of timing out due to the fact that (a) its N_U value was
never dialed down and (b) it uses a single global lock for the
verification step.  I'd thought about improving the performance of the
verification step, but after comparing to the release version, don't
think it's worth running this additional version for any reason.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23130 3a8e244f-b0f2-452b-bcba-4c88e055c3ca