In the spec, spread out the elements of several "one of" lists.
[reviewed by Tom]

I went through the "one of" lists in the grammar productions in the
spec and increased the visual space between those lists' elements, for
improved readability.

I used "$ $ $ $" to add the space in Latex, e.g.

unary-operator: one of
  + - ~ !

==>

unary-operator: one of
  + $ $ $ $ - $ $ $ $ ~ $ $ $ $ !

I don't know who started this approach, I don't think it was me.

This makes the Latex source less readable, but the pdf output more
readable. I propose to accept this trade-off.

I left a couple of lists alone (e.g. "letter" and "digit").
There might be lists that I did not touch because they
did not grep for ": one of".

In the "letter" list, I moved the line break so that all upper-case
letters are on one line and all lower-case letters are on another
line. Visually, it feels slightly better this way.

Reviewer's comment: A more legible approach might be to define a
newcommand and use \hspace in its definition.  Adjusting the spacing
to taste at a later date would then be available at the flick of a
LaTeX length specifier (e.g. 3em).


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23131 3a8e244f-b0f2-452b-bcba-4c88e055c3ca