Minor wordmithing to the definition and description of noinit
[reviewed by lydia and hilde]

These edits are based on concerns from Tom that talking in terms of
assignment may be the wrong thing in terms of a user-defined
assignment operator that, itself, reads fields from a record/class
(say) and will not behave correctly if those fields are uninitialized.
For this reason, changed discussion to talk in terms of reads/writes.
Tightened up a few other things while we were all looking at this.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23132 3a8e244f-b0f2-452b-bcba-4c88e055c3ca