Do not check for perf keys on timeout
[reviewed by lydia] 

Brad noticed that on execution timeout for a performance run we would still
check for the performance keys and more curiously we would print out success
matching performance keys when they clearly weren't matched. 

The reason for this was because on timeout computePerfStats would not return an
error so sub_test assumed key matching happened normally. 

This patch changes things so that we skip perf key checking when we had a
timeout. It also returns an error if the timeout occurred although that error is
now really ignored since we no longer print out success or failure matching
performance keys if there was a timeout (that way you don't get duplicate
errors about not matching perf keys and a timeout.) 



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23151 3a8e244f-b0f2-452b-bcba-4c88e055c3ca