Add memory_order argument to network atomics
[reviewed by Elliot]

This patch adds the memory_order argument with default value to the
network atomics versions of the methods, so that their signatures
match those of the processor atomics.

Note that this is currently dropped on the floor for *all* atomic
types and implementations.

Also, change existing atomics interface tests to pass in the optional
argument, just to test the interface.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23166 3a8e244f-b0f2-452b-bcba-4c88e055c3ca