Added a printout of top CPU consumers upon timedexec's timeout.
[reviewed by Tom]

The motivation is to allow us to see any CPU contenders when a test times out.

TODO: ensure such a printout happens when the compiler times out, too.

The printout command was contributed by Tom, to make sure it works
on the mac, in addition to linux.

I put 'top' before print "timedexec Alarm Clock\n" so that the latter
occurs at the bottom of the test output. Otherwise, IMO, "timedexec
Alarm Clock\n" gets visually lost somewhere in between the test output
and the output of 'top'.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23170 3a8e244f-b0f2-452b-bcba-4c88e055c3ca