Encapsulate ra/ra-cleanloop verification lock code in a record and yield while spin waiting
[reviewed by Greg]

I originally omitted the yield since we expect contention to be low
and that every one will make progress, but it's possible that the lack
of the yield is the cause of timeouts and orphaned processes in the
nightly testing.

While in there, I encapsulated the lock code into a record so the
whole thing looks a lot cleaner.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23182 3a8e244f-b0f2-452b-bcba-4c88e055c3ca