Performance improvement/simplification to serial 1D array iterator
[reviewed by diten]

While working on evaluating nbody performance the other night, I
stumbled across some control flow in the generated code that seemed
unnecessarily complex for what was a fairly simple serial loop over an
array.  It turned out that we have been generating the full strided
range loop idiom generality for 1D array iteration even in the case
that an array is not strided because the array's serial iterator
unnecessarily uses a strided range to implement iteration.  In
rewriting this to avoid the strided range, I saw performance
improvements in the ~10% range for nbody.

This commit is a first baby step toward potentially getting additional
performance out of this case and additional cases working better.
Since my time is limited this week with the pending release and this
patch is working and beneficial, I decided to check it in and then
continue exploring other idioms and approaches after the fact, getting
some updated performance numbers in the nightlies in the meantime2.

The main effect in this patch is to change a loop of the form:

   for i in first..last by step do
     ...

to one of the form:

          var i = first;
          while (i != last) {
            ...
            i += step;
          }

I also removed some explicit calculations against the stride value in
a conditional that already knows the domain/range to be not strided
(i.e., the stride is known to be 1).

The only other change was to update the output of a test that explains
instantiations of ranges, because this change causes us to use fewer
range signatures in simple programs.

As future work, I want to explore the following:

  * switching back to a range for loop that doesn't use strides:

        var i = first;
        for j in 1..dom.dsiNumIndices {
          ...
          i += step;
        }

  * switching my current while loop to use < and/or <=, which David
    says some compilers can do more with

  * moving all of this logic into a walker/bumper iterator in ddata
    which would walk a pointer through memory rather than doing
    calculations on indices

  * applying these techniques to the strided and multidimensional
    cases of the serial iterator; and to apply all of this logic
    to the follower iterator which seems to be written in a more
    naive way (and one that the serial iterator specifically says
    it's avoiding for performance reasons) to see if we can see
    similar benefits in parallel loops (like Stream).

  * correct my apparently improper use of the phrase "beg the
    question" (http://www.qwantz.com/index.php?comic=693)




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23189 3a8e244f-b0f2-452b-bcba-4c88e055c3ca