Commenting out QIO debugging code
[reviewed by diten]

This commit comments out some developer-oriented QIO debugging
code which isn't portable to the Cray C++ compilers when
compiling re2-interface.cc.  To reproduce, load the Cray
PrgEnv-cray module, set CHPL_REGEXP to re2 and build.

Commenting these out rather than looking for a fix in order
to get the release out the door with re2 enabled in as many
environments as possible, and because the following thread
(which exhibits the same symptom) doesn't offer a solution:

  http://stackoverflow.com/questions/14535556/why-doesnt-priu64-work-in-this-code



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23190 3a8e244f-b0f2-452b-bcba-4c88e055c3ca