For qthreads, make the comm task a pthread, not a task.
[Reviewed by Sung.]

In r22839 we changed the comm task from a pthread to a task, in qthreads
tasking.  The idea was that if it knew about the comm task, qthreads
could better manage supply vs. demand for processor resources.  But some
in the group expressed concern that since qthreads uses a cooperative
multitasking model, if a program had enough compute-bound tasks, those
could starve the I/O-bound comm task and prevent forward progress.  We
have now seen indications that this may be happening, so here we change
the comm task back into a pthread.  This pthread is scheduled by the
kernel instead of by qthreads, and the kernel can switch preemptively
when pthreads do not yield cooperatively, so this makes the comm task
more resilient against compute-bound competition.  It can still be
slowed, but it cannot be stopped entirely.

In related and/or enabling work I also did the following:
- removed the code that took the comm task into account when computing
  how many threads to create;
- adjusted the functions that provide access to task private data so
  that they can handle being called from something that isn't actually a
  task (because for "fast" remote forks the comm task will run Chapel
  code that assumes it can ask for information stored in task private
  data);
- adjusted chpl_task_yield() so that it can yield the processor when
  called from something that isn't a task (because the comm task has to
  be able to yield while polling, but cannot be expected to know how the
  tasking layer has chosen to implement it).

While here, I also renamed chapel_qthreads_get_tasklocal() to the more
standard chpl_qthread_get_tasklocal(), removed an excess '_' from the
symbol chpl__qthread_tls_t in a comment, and removed another comment
that had gone out of date.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23192 3a8e244f-b0f2-452b-bcba-4c88e055c3ca