Removing an initializer that assumed CHPL_ATOMICS=intrinsics
[proposed by hilde/mppf, reviewed by vass/bradc]

This file had a C++-style initializer that assumed that CHPL_ATOMICS
was set to intrinsics and therefore that an atomic type could be
initialized with '0'.  This isn't the case with CHPL_ATOMICS=locks
which caused the compile to break for compilers like PGI.

Turns out the initializer wasn't necessary because the reference
count is initialized in the body of the constructor anyway, so
this commit simply removes it.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23193 3a8e244f-b0f2-452b-bcba-4c88e055c3ca