Reorganize #includes to put standard #includes first
[trivial, not reviewed]

This commit reorders the #includes in this file to put all the
standard/system headers first.  For PGI, the old order
caused problems because the symbols that Chapel redefines
(like exit, malloc, etc.) to keep ourselves honest in the
runtime breaks the header files.  This new order permits this
file to be compiled with PGI; though the result isn't that
RE2 works with PGI due to other portability issues in the
re2 distribution itself.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23197 3a8e244f-b0f2-452b-bcba-4c88e055c3ca