Yield while waiting for barriers to satisfy.
[Reviewed by Sung.]

We've been experiencing intermittent hangs in nightly testing with
gasnet+qthreads.  It turns out that at least some of these are due to a
livelock resulting from pthread monopolization in barriers.  In the
runtime startup code we do barrier calls at the beginning and end of the
pre-user-code hook function so that we don't do any pre-user-code work
anywhere until locale 0 has done the module initialization, and we don't
enter the user code (on locale 0) until all locales complete the
pre-user-code hook.  With tasks=qthreads the pre-user-code hook is
called in a Chapel task.  With CHPL_RT_NUM_THREADS_PER_LOCALE=2 we can
get into livelock in this code, though, because the GASNet barrier wait
function does sched_yield() which yields the pthread's claim on the
processor but retains the task's claim on the pthread, rather than
chpl_task_yield() which yields the task's claim on the pthread but keeps
the pthread's claim on the processor (but subject to kernel preemption).
Based on a conversation with Dylan Stark of the qthreads group, qthreads
assigns Chapel tasks to shepherds at the time the associated qthread is
forked.  In the case I studied, the execution order was such that on
locale 1 one task (call it A) was waiting at the entry barrier in the
pre-user-code hook, another task (B) was spinning, waiting for an atomic
flag to become true, and a third task (C), which would set the flag for
which B was waiting, had been queued as a result of a remote executeOn
from locale 0 but was not being started.  There were 2 shepherd threads
and 2 worker threads (1 per shepherd).  I don't know for sure, but I
think the reason for the livelock was that A and C happened to have been
assigned to the same shepherd.  A was monopolizing that shepherd's only
worker by calling sched_yield() instead of chpl_task_yield(), so C
couldn't run.  Meanwhile B was being cooperative with its worker, but C
couldn't take advantage of that B's shepherd (and worker) wasn't the one
C had been assigned to.

To deal with this, make chpl_comm_barrier() call chpl_task_yield() while
waiting for the barrier to be satisfied.

Note that this doesn't by any means solve all the problems we have when
testing gasnet+qthreads.  Especially when the thread count is limited to
2, as it currently is for gasnet+qthreads nightly testing, we get
timeouts, both solid and intermittent, on lots of tests.  However, if I
run without limiting the threads I don't see nearly as many timeouts.
(This is not to say that the performance is then acceptable either,
however.  I haven't checked that.)

The specific file changes are as follows.

runtime/include/chpl-comm.h
  Update comment to say that the chpl_comm_barrier() implementation must
  arrange to do chpl_task_yield() calls while waiting.

runtime/src/comm/gasnet/comm-gasnet.c
  Change chpl_comm_barrier() so that instead of just calling the GASNet
  barrier wait function it loops, trying the barrier and yielding as
  long as it isn't satisifed.  Related to this, change the pre-task-exit
  barrier from its old manually coded try/yield spin implementation to a
  simple call to chpl_comm_barrier(), since that now behaves in the same
  way (minus the nanosleep(), but that wasn't absolutely necessary in
  the old version).

  While here, also get rid of the new barrier_with_id() function.  It's
  an unnecessary complication.  A simpler way to get the same result is
  to use the first byte of the passed-in barrier identification message
  as the ID.  Of course, this does require that all callers to a barrier
  use the same message string.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@23199 3a8e244f-b0f2-452b-bcba-4c88e055c3ca